+++ src/ast/passes/codegen_llvm.h
@@ -56,7 +56,7 @@ public:
   void visit(ExprStatement &expr) override;
   void visit(AssignMapStatement &assignment) override;
   void visit(AssignVarStatement &assignment) override;
-  void visit(If &if_block) override;
+  void visit(If &if_node) override;
   void visit(Unroll &unroll) override;
   void visit(While &while_block) override;
   void visit(For &f) override;
@@ -66,6 +66,7 @@ public:
   void visit(Probe &probe) override;
   void visit(Subprog &subprog) override;
   void visit(Program &program) override;
+  void visit(Block &block) override;
 
   AllocaInst *getHistMapKey(Map &map, Value *log2);
   int getNextIndexForProbe();