+++ src/ast/passes/semantic_analyser.cpp
@@ -220,7 +220,7 @@ AddrSpace SemanticAnalyser::find_addrspace(ProbeType pt)
 void SemanticAnalyser::visit(Builtin &builtin)
 {
   if (builtin.ident == "ctx") {
-    auto probe = get_probe_from_scope(scope_, builtin.loc, builtin.ident);
+    auto probe = get_probe(builtin.loc, builtin.ident);
     if (probe == nullptr)
       return;
     ProbeType pt = probetype(probe->attach_points[0]->provider);
@@ -297,7 +297,7 @@ void SemanticAnalyser::visit(Builtin &builtin)
                                                   "struct task_struct")),
                                  AddrSpace::kernel);
   } else if (builtin.ident == "retval") {
-    auto probe = get_probe_from_scope(scope_, builtin.loc, builtin.ident);
+    auto probe = get_probe(builtin.loc, builtin.ident);
     if (probe == nullptr)
       return;
     ProbeType type = single_provider_type(probe);
@@ -335,7 +335,7 @@ void SemanticAnalyser::visit(Builtin &builtin)
     // Case: @=comm and strncmp(@, "name")
     builtin.type.SetAS(AddrSpace::kernel);
   } else if (builtin.ident == "func") {
-    auto probe = get_probe_from_scope(scope_, builtin.loc, builtin.ident);
+    auto probe = get_probe(builtin.loc, builtin.ident);
     if (probe == nullptr)
       return;
     for (auto *attach_point : probe->attach_points) {
@@ -364,7 +364,7 @@ void SemanticAnalyser::visit(Builtin &builtin)
       }
     }
   } else if (builtin.is_argx()) {
-    auto probe = get_probe_from_scope(scope_, builtin.loc, builtin.ident);
+    auto probe = get_probe(builtin.loc, builtin.ident);
     if (probe == nullptr)
       return;
     ProbeType pt = probetype(probe->attach_points[0]->provider);
@@ -391,7 +391,7 @@ void SemanticAnalyser::visit(Builtin &builtin)
   } else if (!builtin.ident.compare(0, 4, "sarg") &&
              builtin.ident.size() == 5 && builtin.ident.at(4) >= '0' &&
              builtin.ident.at(4) <= '9') {
-    auto probe = get_probe_from_scope(scope_, builtin.loc, builtin.ident);
+    auto probe = get_probe(builtin.loc, builtin.ident);
     if (probe == nullptr)
       return;
     ProbeType pt = probetype(probe->attach_points[0]->provider);
@@ -420,7 +420,7 @@ void SemanticAnalyser::visit(Builtin &builtin)
     builtin.type = CreateUInt64();
     builtin.type.SetAS(addrspace);
   } else if (builtin.ident == "probe") {
-    auto probe = get_probe_from_scope(scope_, builtin.loc, builtin.ident);
+    auto probe = get_probe(builtin.loc, builtin.ident);
     if (probe == nullptr)
       return;
     builtin.type = CreateProbe();
@@ -434,7 +434,7 @@ void SemanticAnalyser::visit(Builtin &builtin)
     }
     builtin.type = CreateUInt32();
   } else if (builtin.ident == "args") {
-    auto probe = get_probe_from_scope(scope_, builtin.loc, builtin.ident);
+    auto probe = get_probe(builtin.loc, builtin.ident);
     if (probe == nullptr)
       return;
     for (auto *attach_point : probe->attach_points) {
@@ -539,7 +539,7 @@ void SemanticAnalyser::visit(Call &call)
     Visit(call.vargs[i]);
   }
 
-  if (auto probe = dynamic_cast<Probe *>(scope_)) {
+  if (auto probe = dynamic_cast<Probe *>(top_level_node_)) {
     for (auto *ap : probe->attach_points) {
       if (!check_available(call, *ap)) {
         LOG(ERROR, call.loc, err_) << call.func << " can not be used with \""
@@ -908,7 +908,7 @@ void SemanticAnalyser::visit(Call &call)
       }
     }
     call.type = CreateUInt64();
-    if (auto probe = dynamic_cast<Probe *>(scope_)) {
+    if (auto probe = dynamic_cast<Probe *>(top_level_node_)) {
       ProbeType pt = single_provider_type(probe);
       // In case of different attach_points, Set the addrspace to none.
       call.type.SetAS(find_addrspace(pt));
@@ -923,7 +923,7 @@ void SemanticAnalyser::visit(Call &call)
     call.type = CreateUInt64();
     call.type.SetAS(AddrSpace::kernel);
   } else if (call.func == "uaddr") {
-    auto probe = get_probe_from_scope(scope_, call.loc, call.func);
+    auto probe = get_probe(call.loc, call.func);
     if (probe == nullptr)
       return;
 
@@ -1166,7 +1166,7 @@ void SemanticAnalyser::visit(Call &call)
           << "signal only accepts string literals or integers";
     }
   } else if (call.func == "path") {
-    auto probe = get_probe_from_scope(scope_, call.loc, call.func);
+    auto probe = get_probe(call.loc, call.func);
     if (probe == nullptr)
       return;
 
@@ -1227,7 +1227,7 @@ void SemanticAnalyser::visit(Call &call)
     }
     call.type = CreateUInt64();
   } else if (call.func == "override") {
-    auto probe = get_probe_from_scope(scope_, call.loc, call.func);
+    auto probe = get_probe(call.loc, call.func);
     if (probe == nullptr)
       return;
 
@@ -1443,11 +1443,9 @@ void SemanticAnalyser::check_stack_call(Call &call, bool kernel)
   call.type = CreateStack(kernel, stack_type);
 }
 
-Probe *SemanticAnalyser::get_probe_from_scope(Scope *scope,
-                                              const location &loc,
-                                              std::string name)
+Probe *SemanticAnalyser::get_probe(const location &loc, std::string name)
 {
-  auto probe = dynamic_cast<Probe *>(scope);
+  auto probe = dynamic_cast<Probe *>(top_level_node_);
   if (probe == nullptr) {
     // Attempting to use probe-specific feature in non-probe context
     if (name.empty()) {
@@ -1523,14 +1521,17 @@ void SemanticAnalyser::visit(Map &map)
 
 void SemanticAnalyser::visit(Variable &var)
 {
-  auto search_val = variable_val_[scope_].find(var.ident);
-  if (search_val != variable_val_[scope_].end()) {
-    var.type = search_val->second;
-  } else {
-    LOG(ERROR, var.loc, err_)
-        << "Undefined or undeclared variable: " << var.ident;
-    var.type = CreateNone();
+  for (auto scope : scope_stack_) {
+    auto search_val = variable_val_[scope].find(var.ident);
+    if (search_val != variable_val_[scope].end()) {
+      var.type = search_val->second;
+      return;
+    }
   }
+
+  LOG(ERROR, var.loc, err_)
+      << "Undefined or undeclared variable: " << var.ident;
+  var.type = CreateNone();
 }
 
 void SemanticAnalyser::visit(ArrayAccess &arr)
@@ -1979,18 +1980,18 @@ void SemanticAnalyser::visit(Ternary &ternary)
   }
 }
 
-void SemanticAnalyser::visit(If &if_block)
+void SemanticAnalyser::visit(If &if_node)
 {
-  Visit(if_block.cond);
+  Visit(if_node.cond);
 
   if (is_final_pass()) {
-    const Type &cond = if_block.cond->type.GetTy();
+    const Type &cond = if_node.cond->type.GetTy();
     if (cond != Type::integer)
-      LOG(ERROR, if_block.loc, err_) << "Invalid condition in if(): " << cond;
+      LOG(ERROR, if_node.loc, err_) << "Invalid condition in if(): " << cond;
   }
 
-  accept_statements(if_block.stmts);
-  accept_statements(if_block.else_stmts);
+  Visit(&if_node.if_block);
+  Visit(&if_node.else_block);
 }
 
 void SemanticAnalyser::visit(Unroll &unroll)
@@ -2011,17 +2012,16 @@ void SemanticAnalyser::visit(Unroll &unroll)
     LOG(ERROR, unroll.loc, err_) << "unroll minimum value is 1";
   }
 
-  accept_statements(unroll.stmts);
+  Visit(&unroll.block);
 }
 
 void SemanticAnalyser::visit(Jump &jump)
 {
   switch (jump.ident) {
     case JumpType::RETURN:
-      if (jump.return_value) {
+      if (jump.return_value)
         Visit(jump.return_value);
-      }
-      if (auto subprog = dynamic_cast<Subprog *>(scope_)) {
+      if (auto subprog = dynamic_cast<Subprog *>(top_level_node_)) {
         if ((subprog->return_type.IsVoidTy() !=
              (jump.return_value == nullptr)) ||
             (jump.return_value &&
@@ -2054,7 +2054,7 @@ void SemanticAnalyser::visit(While &while_block)
   Visit(while_block.cond);
 
   loop_depth_++;
-  accept_statements(while_block.stmts);
+  Visit(&while_block.block);
   loop_depth_--;
 }
 
@@ -2151,9 +2151,12 @@ void SemanticAnalyser::visit(For &f)
 
   // Validate decl
   const auto &decl_name = f.decl->ident;
-  if (variable_val_[scope_].find(decl_name) != variable_val_[scope_].end()) {
-    LOG(ERROR, f.decl->loc, err_)
-        << "Loop declaration shadows existing variable: " + decl_name;
+
+  for (auto scope : scope_stack_) {
+    if (variable_val_[scope].find(decl_name) != variable_val_[scope].end()) {
+      LOG(ERROR, f.decl->loc, err_)
+          << "Loop declaration shadows existing variable: " + decl_name;
+    }
   }
 
   // Validate expr
@@ -2190,27 +2193,25 @@ void SemanticAnalyser::visit(For &f)
   // and declared before the loop. These will be passed into the loop callback
   // function as the context parameter.
   CollectNodes<Variable> vars_referenced;
-  std::unordered_set<std::string> var_set;
+  std::unordered_set<std::string> found_vars;
   for (auto *stmt : f.stmts) {
-    const auto &live_vars = variable_val_[scope_];
-    vars_referenced.run(*stmt, [&live_vars, &var_set](const auto &var) {
-      if (live_vars.find(var.ident) == live_vars.end())
+    vars_referenced.run(*stmt, [this, &f, &found_vars](const auto &var) {
+      if (found_vars.find(var.ident) != found_vars.end())
         return false;
-      if (var_set.find(var.ident) != var_set.end())
-        return false;
-      var_set.insert(var.ident);
-      return true;
-    });
-  }
 
-  // Collect a list of variables which are used in the loop without having been
-  // used before. This is a hack to simulate block scoping in the absence of the
-  // real thing (#3017).
-  CollectNodes<Variable> new_vars;
-  for (auto *stmt : f.stmts) {
-    const auto &live_vars = variable_val_[scope_];
-    new_vars.run(*stmt, [&live_vars](const auto &var) {
-      return live_vars.find(var.ident) == live_vars.end();
+      bool found = false;
+      for (auto scope : scope_stack_) {
+        const auto &live_vars = variable_val_[scope];
+        if (live_vars.find(var.ident) != live_vars.end()) {
+          found = true;
+          break;
+        }
+      }
+      if (!found) {
+        return false;
+      }
+      found_vars.insert(var.ident);
+      return true;
     });
   }
 
@@ -2235,11 +2236,15 @@ void SemanticAnalyser::visit(For &f)
   }
   f.decl->type = CreateTuple(bpftrace_.structs.AddTuple({ keytype, *mapval }));
 
-  variable_val_[scope_][decl_name] = f.decl->type;
+  scope_stack_.push_back(&f);
+
+  variable_val_[scope_stack_.back()][decl_name] = f.decl->type;
 
   loop_depth_++;
   accept_statements(f.stmts);
   loop_depth_--;
+
+  scope_stack_.pop_back();
 
   // Currently, we do not pass BPF context to the callback so disable builtins
   // which require ctx access.
@@ -2253,14 +2258,6 @@ void SemanticAnalyser::visit(For &f)
       LOG(ERROR, builtin.loc, err_)
           << "'" << builtin.ident << "' builtin is not allowed in a for-loop";
     }
-  }
-
-  // Decl variable is not valid beyond this for-loop
-  variable_val_[scope_].erase(decl_name);
-
-  // Variables declared in a for-loop are not valid beyond it
-  for (const Variable &var : new_vars.nodes()) {
-    variable_val_[scope_].erase(var.ident);
   }
 
   // Finally, create the context tuple now that all variables inside the loop
@@ -2306,7 +2303,7 @@ void SemanticAnalyser::visit(FieldAccess &acc)
   }
 
   if (type.is_funcarg) {
-    auto probe = get_probe_from_scope(scope_, acc.loc);
+    auto probe = get_probe(acc.loc);
     if (probe == nullptr)
       return;
     auto arg = bpftrace_.structs.GetProbeArg(*probe, acc.field);
@@ -2365,7 +2362,7 @@ void SemanticAnalyser::visit(FieldAccess &acc)
   std::map<std::string, std::weak_ptr<const Struct>> structs;
 
   if (type.is_tparg) {
-    auto probe = get_probe_from_scope(scope_, acc.loc);
+    auto probe = get_probe(acc.loc);
     if (probe == nullptr)
       return;
 
@@ -2505,7 +2502,7 @@ void SemanticAnalyser::visit(Cast &cast)
   // case : BEGIN { @foo = (struct Foo)0; }
   // case : profile:hz:99 $task = (struct task_struct *)curtask.
   if (cast.type.GetAS() == AddrSpace::none) {
-    if (auto probe = dynamic_cast<Probe *>(scope_)) {
+    if (auto probe = dynamic_cast<Probe *>(top_level_node_)) {
       ProbeType type = single_provider_type(probe);
       cast.type.SetAS(find_addrspace(type));
     } else {
@@ -2619,33 +2616,44 @@ void SemanticAnalyser::visit(AssignVarStatement &assignment)
   Visit(assignment.expr);
 
   std::string var_ident = assignment.var->ident;
-  auto search = variable_val_[scope_].find(var_ident);
 
   auto &assignTy = assignment.expr->type;
 
-  if (search != variable_val_[scope_].end()) {
-    if (search->second.IsNoneTy()) {
-      if (is_final_pass()) {
-        LOG(ERROR, assignment.loc, err_) << "Undefined variable: " + var_ident;
-      } else {
-        search->second = assignTy;
+  Node *var_scope = nullptr;
+  for (auto scope : scope_stack_) {
+    auto search = variable_val_[scope].find(var_ident);
+    if (search != variable_val_[scope].end()) {
+      if (search->second.IsNoneTy()) {
+        if (is_final_pass()) {
+          LOG(ERROR, assignment.loc, err_)
+              << "Undefined variable: " + var_ident;
+          return;
+        } else {
+          search->second = assignTy;
+        }
+      } else if ((search->second.IsStringTy() && assignTy.IsStringTy()) ||
+                 (search->second.IsTupleTy() && assignTy.IsTupleTy())) {
+        update_string_size(search->second, assignTy);
+      } else if (!search->second.IsSameType(assignTy)) {
+        LOG(ERROR, assignment.loc, err_)
+            << "Type mismatch for " << var_ident << ": "
+            << "trying to assign value of type '" << assignTy
+            << "' when variable already contains a value of type '"
+            << search->second << "'";
+        return;
       }
-    } else if ((search->second.IsStringTy() && assignTy.IsStringTy()) ||
-               (search->second.IsTupleTy() && assignTy.IsTupleTy())) {
-      update_string_size(search->second, assignTy);
-    } else if (!search->second.IsSameType(assignTy)) {
-      LOG(ERROR, assignment.loc, err_)
-          << "Type mismatch for " << var_ident << ": "
-          << "trying to assign value of type '" << assignTy
-          << "' when variable already contains a value of type '"
-          << search->second << "'";
+      var_scope = scope;
+      break;
     }
-  } else {
-    // This variable hasn't been seen before
-    variable_val_[scope_].insert({ var_ident, assignment.expr->type });
   }
 
-  auto &storedTy = variable_val_[scope_][var_ident];
+  if (var_scope == nullptr) {
+    variable_val_[scope_stack_.back()].insert(
+        { var_ident, assignment.expr->type });
+    var_scope = scope_stack_.back();
+  }
+
+  auto &storedTy = variable_val_[var_scope][var_ident];
 
   assignment.var->type = storedTy;
 
@@ -2999,11 +3007,17 @@ void SemanticAnalyser::visit(AttachPoint &ap)
   }
 }
 
+void SemanticAnalyser::visit(Block &block)
+{
+  scope_stack_.push_back(&block);
+  accept_statements(block.stmts);
+  scope_stack_.pop_back();
+}
+
 void SemanticAnalyser::visit(Probe &probe)
 {
   auto aps = probe.attach_points.size();
-
-  scope_ = &probe;
+  top_level_node_ = &probe;
 
   for (AttachPoint *ap : probe.attach_points) {
     if (!listing_ && aps > 1 && ap->provider == "iter") {
@@ -3015,9 +3029,7 @@ void SemanticAnalyser::visit(Probe &probe)
   if (probe.pred) {
     Visit(probe.pred);
   }
-  for (Statement *stmt : probe.stmts) {
-    Visit(stmt);
-  }
+  Visit(&probe.block);
 }
 
 void SemanticAnalyser::visit(Config &config)
@@ -3027,11 +3039,13 @@ void SemanticAnalyser::visit(Config &config)
 
 void SemanticAnalyser::visit(Subprog &subprog)
 {
-  scope_ = &subprog;
+  scope_stack_.push_back(&subprog);
+  top_level_node_ = &subprog;
   for (SubprogArg *arg : subprog.args) {
-    variable_val_[scope_].insert({ arg->name(), arg->type });
+    variable_val_[scope_stack_.back()].insert({ arg->name(), arg->type });
   }
   Visitor::visit(subprog);
+  scope_stack_.pop_back();
 }
 
 void SemanticAnalyser::visit(Program &program)